requirements.txt
fastapi
uvicorn
sentence-transformers
psycopg2-binary
pgvector
python-dotenv
groq
langchain-text-splitters
pydantic
